requirement.txt
torch>=2.0.0
torchvision>=0.15.0
transformers==4.41.2
tokenizers==0.19.1
pillow>=9.0.0
opencv-python>=4.7.0
matplotlib>=3.5.0
tqdm>=4.65.0
kornia>=0.6.0
numpy<2.0.0,>=1.24.0
scipy>=1.11.0